% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Cho2016-data.R
\docType{data}
\name{Cho2016}
\alias{Cho2016}
\title{Cho's (2016) artificial unidimensional data of four items}
\format{
An object of class \code{matrix} (inherits from \code{array}) with 4 rows and 4 columns.
}
\source{
Cho, E. (2016). Making reliability reliable: A systematic approach to
reliability coefficients. Organizational Research Methods, 19(4), 651-682.
}
\usage{
data(Cho2016)
}
\description{
A hypothetical data consisting of four items.
}
\author{
Eunseong Cho, \email{bene@kw.ac.kr}
}
\keyword{datasets}
