% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ChoA-data.R
\docType{data}
\name{ChoA}
\alias{ChoA}
\title{Cho's (in press) artificial unidimensional data of three items}
\format{
An object of class \code{matrix} (inherits from \code{array}) with 3 rows and 3 columns.
}
\source{
Cho, E. (in press). Neither Cronbach's alpha nor McDonald's omega: A 
comment on Sijtsma and Pfadt. Psychometrika.
}
\usage{
data(ChoA)
}
\description{
A hypothetical data consisting of three items. Item 1 has a variance of 1, 
item 2 is a copy of item 1, and item 3 is item 1 multiplied by -1. This  data
show that coefficient alpha derives negative reliability.
}
\author{
Eunseong Cho, \email{bene@kw.ac.kr}
}
\keyword{datasets}
