% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ChoKim-data.R
\docType{data}
\name{ChoKim}
\alias{ChoKim}
\title{Cho and Kim's (2015) artificial unidimensional data of five items}
\format{
An object of class \code{matrix} (inherits from \code{array}) with 5 rows and 5 columns.
}
\source{
Cho, E., & Kim, S. (2015). Cronbach’s coefficient alpha: Well known
but poorly understood. Organizational Research Methods, 18(2), 207-230.
}
\usage{
data(ChoKim)
}
\description{
A hypothetical data consisting of five items.
}
\author{
Eunseong Cho, \email{bene@kw.ac.kr}
}
\keyword{datasets}
