% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correlated_factors.R
\name{correlated_factors}
\alias{correlated_factors}
\title{Obtain correlated factors reliability estimates of two-dimensional data}
\usage{
correlated_factors(x, until, nonneg_loading = FALSE)
}
\arguments{
\item{x}{observed item scores or their covariances}

\item{until}{The number of items up to the first sub-construct}

\item{nonneg_loading}{if TRUE, constraint loadings to nonnegative values}
}
\value{
a correlated factors reliability estimate
}
\description{
Obtain correlated factors reliability estimates of two-dimensional data. It is a kind
of multidimensional CFA reliability, and it is the reliability derived from
the correlated factors model. Items should be grouped by each sub-dimension. For
example, the first dimension consists of the first item through the until item.
}
\examples{
correlated_factors(Osburn_moderate, 4)
}
\references{
Cho, E. (2016). Making reliability reliable: A systematic
approach to reliability coefficients. Organizational Research Methods, 19(4),
651-682. https://doi.org/10.1177/1094428116656239
}
\author{
Eunseong Cho, \email{bene@kw.ac.kr}
}
