% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DeLeeuw-data.R
\docType{data}
\name{DeLeeuw}
\alias{DeLeeuw}
\title{De Leeuw's (1983) unidimensional data}
\format{
An object of class \code{matrix} (inherits from \code{array}) with 6 rows and 6 columns.
}
\source{
De Leeuw, J. (1983). Models and methods for the analysis of 
correlation coefficients. Journal of Econometrics, 22,113-137.
}
\usage{
data(DeLeeuw)
}
\description{
De Leeuw's (1983) political survey of six items. This dataset is from Ten 
Berge and Socan (2004).
}
\references{
Ten Berge, J. M. F., & Sočan, G. (2004). The greatest lower bound to 
the reliability of a test and the hypothesis of unidimensionality. 
Psychometrika, 69(4), 613-625.
}
\author{
Eunseong Cho, \email{bene@kw.ac.kr}
}
\keyword{datasets}
