% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FeldtBrennan-data.R
\docType{data}
\name{FeldtBrennan}
\alias{FeldtBrennan}
\title{Feldt and Brennan's (1989) unidimensional data}
\format{
An object of class \code{matrix} (inherits from \code{array}) with 4 rows and 4 columns.
}
\source{
Feldt, L. S., & Brennan, R. L. (1989). Reliability. In R. L. Linn 
(Ed.), Educational measurement (3rd ed., pp. 105-146). American Council on 
Education and Macmillan.
}
\usage{
data(FeldtBrennan)
}
\description{
Feldt and Brennan's (1989) artificial unidimensional dataset consisting of 4 items.
}
\author{
Eunseong Cho, \email{bene@kw.ac.kr}
}
\keyword{datasets}
