% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_cov.r
\name{get_cov}
\alias{get_cov}
\title{Obtain the covariance matrix}
\usage{
get_cov(x, cor = FALSE)
}
\arguments{
\item{x}{A dataframe or a matrix}

\item{cor}{if TRUE, return correlation matrix. if FALSE, return covariance matrix}
}
\value{
The covariance or correlation matrix
}
\description{
If the input data is a square matrix, it is converted into a matrix,
otherwise the covariance matrix is obtained.
}
