% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gilmer.r
\name{gilmer}
\alias{gilmer}
\title{Obtain the Gilmer-Feldt reliability coefficient}
\usage{
gilmer(x)
}
\arguments{
\item{x}{A dataframe or a matrix (unidimensional)}
}
\value{
The Gilmer-Fedlt coefficient
}
\description{
It is a unidimensional reliability coefficient
based on a congeneric model. The congeneric model is a model that allows
the length, discrimination, or importance of items to be different,
and is the least restrictive model among the models derived from
the classical test theory. The Gilmer-Feldt coefficient has the advantage
of being less computational than congeneric reliability (Joreskog 1971)
which uses confirmatory factor analysis.However, the Gilmer-Feldt coefficient
 derives a value very close to congeneric reliability (Cho in press).
 Feldt and Charter (2003) offers a user-friendly review of the Gilmer-Feldt
 coefficient.
}
\examples{
gilmer(Graham1)
}
\references{
Cho, E. (in press). Neither Cronbach's alpha nor McDonald's omega: A comment on Sijtsma and Pfadt. Psychometrika.

Feldt, L. S., & Charter, R. A. (2003). Estimation of internal consistency reliability when test parts vary in effective length. Measurement and Evaluation in Counseling and Development, 36(1), 23-27

Gilmer, J. S., & Feldt, L. S. (1983). Reliability estimation for a test with parts of unknown lengths. Psychometrika, 48(1), 99–111.

Jöreskog, K. G. (1971). Statistical analysis of sets of congeneric tests. Psychometrika, 36(2), 109–133.
}
