% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Graham1-data.R
\docType{data}
\name{Graham1}
\alias{Graham1}
\title{Graham's (2006) first unidimensional data}
\format{
An object of class \code{matrix} (inherits from \code{array}) with 5 rows and 5 columns.
}
\source{
Graham, J. M. (2006). Congeneric and (essentially) tau-equivalent 
estimates of score reliability what they are and how to use them. Educational 
and Psychological Measurement, 66(6), 930-944.
Education and Macmillan.
}
\usage{
data(Graham1)
}
\description{
Graham's (2006) artificial dataset consisting of five items.This is the 
second dataset in his Table 2. The last item is about five times as important 
as the other items. This dataset provides a good indication of how sensitive 
each reliability coefficient is to the violation of the tau equivalence 
assumption. The CFA reliability coefficient presented by the author is .97.
}
\author{
Eunseong Cho, \email{bene@kw.ac.kr}
}
\keyword{datasets}
