% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Graham2-data.R
\docType{data}
\name{Graham2}
\alias{Graham2}
\title{Graham's (2006) second unidimensional data}
\format{
An object of class \code{matrix} (inherits from \code{array}) with 4 rows and 4 columns.
}
\source{
Graham, J. M., & Conoley, C. W. (2006). The role of marital 
attributions in the relationship between life stressors and marital quality. 
Personal Relationships, 13(2), 231-241.
}
\usage{
data(Graham2)
}
\description{
Graham's (2006) actual dataset consisting of four items.This dataset is 
originally from Graham and Conoley (2006) and is a part of the measurement 
of the Dyadic Adjustment Scale. This dataset appears in Table 3. This dataset 
violates the assumption of essential tau-equivalence, and the first and 
second items are much more important than the third and fourth items. The CFA 
reliability coefficient presented by the author is .83.
}
\references{
Graham, J. M. (2006). Congeneric and (essentially) tau-equivalent 
estimates of score reliability what they are and how to use them. Educational 
and Psychological Measurement, 66(6), 930-944.
Education and Macmillan.
}
\author{
Eunseong Cho, \email{bene@kw.ac.kr}
}
\keyword{datasets}
