% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hancock.R
\name{hancock}
\alias{hancock}
\title{Obtain Hancock's H (CFA version of maximal reliability)}
\usage{
hancock(x, nonneg_loading = FALSE)
}
\arguments{
\item{x}{a dataframe or a cov (unidimensional)}

\item{nonneg_loading}{if TRUE, constraint loadings to nonnegative values}
}
\value{
Hancock's H
}
\description{
It is the confirmatory factor analysis (CFA) version of maximal
reliability. This coefficient takes the standardized factor loading as the
reliability of each item, and finds the weight that maximizes the
reliability. Hence, Hancock's H shows a different result than the reliability
estimator using conventional unit weights.
}
\examples{
hancock(Graham1)
}
\references{
Cho, E. (in press). Neither Cronbach's alpha nor McDonald's
omega: A comment on Sijtsma and Pfadt. Psychometrika.

Hancock, G., & Mueller, R. O. (2001). Rethinking construct
reliability within latent variable systems. In R. Cudeck, S. du Toit, & D.
Sörbom (Eds.), Structural equation modeling: Present and future-A festschrift
in honor of Karl Jöreskog (pp. 195-216). Scientific Software International.

Li, H., Rosenthal, R., & Rubin, D. B. (1996). Reliability of
measurement in psychology: From Spearman-Brown to maximal reliability.
Psychological Methods, 1(1), 98-107.

McNeish, D. (2017). Thanks coefficient alpha, we’ll take it from
here. Psychological Methods, 23(3), 412-433.
}
