% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hancockan.R
\name{hancockan}
\alias{hancockan}
\title{Obtain Ten Berge and Zegers (2004) mu3}
\usage{
hancockan(x)
}
\arguments{
\item{x}{a dataframe or a matrix (unidimensional)}
}
\description{
Obtain Ten Berge and Zegers (2004) mu3.
}
\details{
Hancock and An (2020) published a reliability estimator they called the
"closed-form omega", which is designed to approximate a unidimensional
confirmatory factor analysis reliability estimator (i.e., joreskog).
}
\examples{
hancockan(Graham1)
}
\references{
Hancock, G. R., & An, J. (2020). A Closed-Form Alternative for
Estimating ω Reliability under Unidimensionality. Measurement:
Interdisciplinary Research and Perspectives, 18(1), 1-14.
https://doi.org/10.1080/15366367.2019.1656049
}
\author{
Eunseong Cho, \email{bene@kw.ac.kr}
}
