% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HancockMueller-data.R
\docType{data}
\name{HancockMueller}
\alias{HancockMueller}
\title{Hancock and Mueller's (2001)  unidimensional data of four items}
\format{
An object of class \code{matrix} (inherits from \code{array}) with 4 rows and 4 columns.
}
\source{
Hancock, G., & Mueller, R. O. (2001). Rethinking construct 
reliability within latent variable systems. In R. Cudeck, S. du Toit, & D. 
Sörbom (Eds.), Structural equation modeling: Present and future—A festschrift 
in honor of Karl Jöreskog (pp. 195–216). Scientific Software International.
}
\usage{
data(HancockMueller)
}
\description{
Hancock and Mueller's artificial unidimensional dataset consisting of four 
items.
}
\author{
Eunseong Cho, \email{bene@kw.ac.kr}
}
\keyword{datasets}
