% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heise.R
\name{heise}
\alias{heise}
\title{Obtain Heise-Bohrnstedt's Omega (another factor analysis reliability)}
\usage{
heise(x)
}
\arguments{
\item{x}{a dataframe or a matrix (unidimensional)}
}
\value{
Heise-Bohrnstedt's Omega
}
\description{
Heise-Bohrnstedt's (1970) Omega is an factor analysis (FA) reliability.
This formula is different from the FA reliability we use today and yields a
larger value (Cho, in press). McDonald (1999) referred to all FA reliability
as omega, and capitalized omega was used to distinguish it from McDonald's
omega.
}
\examples{
heise(Graham1)
}
\references{
Cho, E. (in press). Neither Cronbach's alpha nor McDonald's
omega: A comment on Sijtsma and Pfadt. Psychometrika.

Heise, D. R., & Bohrnstedt, G. W. (1970). Validity, invalidity,
and reliability. Sociological Methodology, 2, 104-129.

McDonald, R. P. (1999). Test theory: A unified treatment.
Lawrence Erlbaum.
}
