% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HeiseBohrnstedt-data.R
\docType{data}
\name{HeiseBohrnstedt}
\alias{HeiseBohrnstedt}
\title{Heise and Bohrnstedt's (1970) unidimensional data}
\format{
An object of class \code{matrix} (inherits from \code{array}) with 6 rows and 6 columns.
}
\source{
Heise, D. R., & Bohrnstedt, G. W. (1970). Validity, invalidity, and 
reliability. Sociological Methodology, 2, 104-129.
}
\usage{
data(HeiseBohrnstedt)
}
\description{
Heise and Bohrnstedt's (1970) unidimensional data consisting of six items. 
The peculiarity of this dataset is that Guttman's lambda5 and lambda6 produce 
smaller reliability estimates than coefficient alpha (Cho in press).
}
\references{
Cho, E. (in press). Neither Cronbach's alpha nor McDonald's omega: 
A comment on Sijtsma and Pfadt. Psychometrika.
Education and Macmillan.
}
\author{
Eunseong Cho, \email{bene@kw.ac.kr}
}
\keyword{datasets}
