% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/joreskog.R
\name{joreskog}
\alias{joreskog}
\title{Obtain Jöreskog's congeneric reliability (Unidimensional CFA reliability)}
\usage{
joreskog(x, nonneg_loading = FALSE)
}
\arguments{
\item{x}{a dataframe or a matrix (unidimensional)}

\item{nonneg_loading}{if TRUE, constraint loadings to nonnegative values}
}
\value{
congeneric reliability coefficient
}
\description{
Congeneric reliability is a reliability coefficient derived from
unidimensional confirmatory factor analysis (CFA).
}
\details{
Features: Congeneric reliability is a unidimensional reliability
coefficient based on a unidimensional confirmatory factor analysis (CFA)
model.

Name: Congeneric reliability is called by a variety of names, general users
usually call it composite reliability, and reliability researchers often call
it omega. One of the reasons for this confusion is that studies that first
proposed this coefficient (Jöreskog 1971) did not give this formula a name
 (Cho 2016). Jöreskog (1971) proposed a matrix-form formula, and the
 commonly known non-matrix formula appears in Werts et al. (1974).

Frequency of use: Congeneric reliability is the second most commonly used
reliability coefficient after coefficient alpha (Cho 2016)

Accuracy: Congeneric reliability is the most accurate reliability coefficient
along with the Feldt-Gilmer coefficient (Cho in press)

Computation: This function uses maximum likelihood as estimation,
unstandardized covariance matrix as input, and lavaan package as software.
}
\examples{
joreskog(Graham1)
}
\references{
Cho, E. (2016). Making reliability reliable: A systematic
approach to reliability coefficients. Organizational Research Methods, 19(4),
651-682.

Cho, E. (in press). Neither Cronbach's alpha nor McDonald's
omega: A comment on Sijtsma and Pfadt. Psychometrika.

Jöreskog, K. G. (1971). Statistical analysis of sets of
congeneric tests. Psychometrika, 36(2), 109-133.

Werts, C. E., Linn, R. L., & Jöreskog, K. G. (1974). Intraclass
reliability estimates: Testing structural assumptions. Educational and
Psychological Measurement, 34, 25-33.
}
\seealso{
[gilmer()] for the Gilmer-Feldt coefficient

[feldt()] for classical congeneric reliability coefficient

[psych::omega()] for a related function of the package psych

[MBESS::ci.reliability()] for a related function of the package MBESS

[Lambda4::omega.tot()] for a related function of the package Lambda4
}
