% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kaisercaffrey.R
\name{kaisercaffrey}
\alias{kaisercaffrey}
\title{Obtain Kaiser-Caffrey's alpha (principal component analysis reliability)}
\usage{
kaisercaffrey(x, print = TRUE)
}
\arguments{
\item{x}{a dataframe or a matrix (unidimensional)}

\item{print}{If TRUE, the result is printed to the screen.}
}
\value{
Kaiser-Caffrey's alpha
}
\description{
Kaiser-Caffrey's (1965) alpha is the principal component analysis (PCA)
reliability. They presented this formula in the context of factor analysis,
but Bentler (1968) showed that it was in fact PCA reliability. Armor (1974),
citing Bentler (1968), referred to this formula as theta, and some studies
refer to it as Armor's theta. Kaiser and Caffrey (1965) labeled this formula
alpha, and people may have mistaken it for coefficient alpha. See Vehkalahti
(2000) and Cho(in press) for further explanation of this formula.
}
\examples{
kaisercaffrey(Graham1)
}
\references{
Armor, D. J. (1974). Theta reliability and factor scaling.
In H. L. Costner (Ed.), Sociological methodology (pp. 17-50). Jossey-Bass.

Bentler, P. M. (1968). Alpha-maximized factor analysis (alphamax)
: Its relation to alpha and canonical factor analysis. Psychometrika, 33(3),
335-345.

Cho, E. (in press). Neither Cronbach's alpha nor McDonald's
omega: A comment on Sijtsma and Pfadt. Psychometrika.

Kaiser, H. F., & Caffrey, J. (1965). Alpha factor analysis.
Psychometrika, 30(1), 1-14.

Vehkalahti, K. (2000). Reliability of measurement scales:
Tarkkonen's general method supersedes Cronbach's alpha. University of
Helsinki.
}
