% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Kristof-data.R
\docType{data}
\name{Kristof}
\alias{Kristof}
\title{Kristof's (1971) unidimensional data}
\format{
An object of class \code{matrix} (inherits from \code{array}) with 3 rows and 3 columns.
}
\usage{
data(Kristof)
}
\description{
Kristof's (1971) unidimensional data consisting of three items.This dataset
is the result of the Scholastic Aptitude Test and is taken from Joreskog (1978).
}
\references{
Kristof, W. On the theory of a set of tests which differ only in
length. Psychometrika, 1971, 36, 207-225.

Jöreskog, K. G. (1978). Structural analysis of covariance and
correlation matrices. Psychometrika, 43(4), 443-477.
}
\author{
Eunseong Cho, \email{bene@kw.ac.kr}
}
\keyword{datasets}
