% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Lord-data.R
\docType{data}
\name{Lord}
\alias{Lord}
\title{Lord's (1957) unidimensional data}
\format{
An object of class \code{matrix} (inherits from \code{array}) with 4 rows and 4 columns.
}
\source{
Lord, F. M.(1957). A significance test for the hypothesis that two 
variables measure the same trait except for errors of measurement. 
Psychometrika, 22, 207-220
}
\usage{
data(Lord)
}
\description{
Lord's (1957) unidimensional data consisting of four items.
}
\references{
Jöreskog, K. G. (1971). Statistical analysis of sets of 
congeneric tests. Psychometrika, 36(2), 109-133.
}
\author{
Eunseong Cho, \email{bene@kw.ac.kr}
}
\keyword{datasets}
