% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LordNovick-data.R
\docType{data}
\name{LordNovick}
\alias{LordNovick}
\title{Lord and Novick's (1968)  unidimensional data of four items}
\format{
An object of class \code{matrix} (inherits from \code{array}) with 4 rows and 4 columns.
}
\source{
Lord, F. M., & Novick, M. R. (1968). Statistical theories of mental 
test scores. Addison-Wesley.
}
\usage{
data(LordNovick)
}
\description{
A unidimensional dataset consisting of four items. This is the result of 
the Educational Testing Service's Test of English as a Foreign Language 
(TOEFL). Lord and Novick (1968) p.91. This dataset was used by Bentler and 
Woodward (1980) and Revelle and Zinbarge (2009).
}
\references{
Bentler, P. M., and J. Arthur Woodward. 1980. Inequalities among 
Lower Bounds to Reliability: With Applications to Test Construction and 
Factor Analysis. Psychometrika 45(2):249-67.

Revelle, William, and Richard E. Zinbarg. 2009. Coefficients 
Alpha, Beta, Omega, and the Glb: Comments on Sijtsma. Psychometrika 74(1):
145-54
}
\author{
Eunseong Cho, \email{bene@kw.ac.kr}
}
\keyword{datasets}
