% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maximal_reliability.R
\name{maximal_reliability}
\alias{maximal_reliability}
\title{Obtain maximal reliability estimates of two-dimensional data}
\usage{
maximal_reliability(x, until)
}
\arguments{
\item{x}{observed item scores or their covariances}

\item{until}{The number of items up to the first sub-construct}
}
\value{
a maximal reliability estimate
}
\description{
Obtain maximal reliability estimates of two-dimensional data. Items should be
grouped by each sub-dimension. For example, the first dimension consists of
the first item through the until item.
}
\examples{
maximal_reliability(Osburn_moderate, 4)
maximal_reliability(Sijtsma2a, c(2, 4))
}
\references{
Li, H., Rosenthal, R., & Rubin, D. B. (1996). Reliability of
measurement in psychology: From Spearman-Brown to maximal reliability.
Psychological Methods, 1(1), 98-107. https://doi.org/10.1037/1082-989X.1.1.98
}
\author{
Eunseong Cho, \email{bene@kw.ac.kr}
}
