% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/McDonald1-data.R
\docType{data}
\name{McDonald1}
\alias{McDonald1}
\title{Mcdonald's (1999)  unidimensional data of five items}
\format{
An object of class \code{matrix} (inherits from \code{array}) with 5 rows and 5 columns.
}
\source{
McDonald, R. P. (1999). Test theory: A unified treatment. 
Lawrence Erlbaum.
}
\usage{
data(McDonald1)
}
\description{
A Satisfaction with Life Scale dataset consisting of five items.Table 6.3.
}
\author{
Eunseong Cho, \email{bene@kw.ac.kr}
}
\keyword{datasets}
