% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Miller-data.R
\docType{data}
\name{Miller}
\alias{Miller}
\title{Miller's (1995) artificial unidimensional data}
\format{
An object of class \code{matrix} (inherits from \code{array}) with 4 rows and 4 columns.
}
\source{
Miller, M. B. (1995). Coefficient alpha: A basic introduction from 
the perspectives of classical test theory and structural equation modeling. 
Structural Equation Modeling: A Multidisciplinary Journal, 2(3), 255-273.
}
\usage{
data(Miller)
}
\description{
Miller's (1995) artificial unidimensional data consisting of four items.
}
\author{
Eunseong Cho, \email{bene@kw.ac.kr}
}
\keyword{datasets}
