% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mu1.R
\name{mu1}
\alias{mu1}
\title{Obtain Ten Berge and Zegers' (1978) mu1}
\usage{
mu1(x, print = TRUE)
}
\arguments{
\item{x}{a dataframe or a matrix (unidimensional)}

\item{print}{If TRUE, the result is printed to the screen.}
}
\description{
Obtain Ten Berge and Zegers' (1978) mu1. mu1 equals Guttman's (1945) lambda2.
}
\references{
Guttman, L. (1945). A basis for analyzing test-retest reliability.
 Psychometrika, 10(4), 255-282.

Ten Berge, J. M. F., & Zegers, F. E. (1978). A series of lower
bounds to the reliability of a test. Psychometrika, 43(4), 575-579.
}
\author{
Eunseong Cho, \email{bene@kw.ac.kr}
}
