% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mu2.R
\name{mu2}
\alias{mu2}
\title{Obtain Ten Berge and Zegers' (1978) mu2}
\usage{
mu2(x, print = TRUE)
}
\arguments{
\item{x}{a dataframe or a matrix (unidimensional)}

\item{print}{If TRUE, the result is printed to the screen.}
}
\description{
Obtain Ten Berge and Zegers' (1978) mu2.
}
\examples{
mu2(Graham1)
}
\references{
Ten Berge, J. M. F., & Zegers, F. E. (1978). A series of lower
bounds to the reliability of a test. Psychometrika, 43(4), 575-579.
}
\author{
Eunseong Cho, \email{bene@kw.ac.kr}
}
