% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mu4.R
\name{mu4}
\alias{mu4}
\title{Obtain Ten Berge and Zegers (1978) mu4}
\usage{
mu4(x)
}
\arguments{
\item{x}{a dataframe or a matrix (unidimensional)}
}
\description{
Obtain Ten Berge and Zegers (1978) mu4.
}
\details{
The original formula and the formula of psych's tenberge() are different.
There is a high possibility that the original formula is incorrect and psych's
version is correct. According to Equation (4) of the original article, mu
should increase monotonically (e.g., mu4>=mu3), but if the original formula is
followed, it may decrease in some cases. The formula of the original paper is
2h, but changing it to 2^h solves this problem. This function follow the latter
interpretation.
}
\examples{
mu4(Graham1)
}
\references{
Ten Berge, J. M. F., & Zegers, F. E. (1978). A series of lower
bounds to the reliability of a test. Psychometrika, 43(4), 575-579.
}
\author{
Eunseong Cho, \email{bene@kw.ac.kr}
}
