% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multi_parallel.R
\name{multi_parallel}
\alias{multi_parallel}
\title{Obtain multidimensional parallel reliability}
\usage{
multi_parallel(x, until)
}
\arguments{
\item{x}{observed item scores or their covariances}

\item{until}{The number of items up to the first sub-construct}
}
\value{
a multidimensional parallel reliability estimate
}
\description{
Multidimensional parallel reliability is derived from the multidimensional parallel model (Cho, 2016).
This is equivalent to entering the correlation instead of the covariance into the stratified alpha formula.
}
\examples{
multi_parallel(Osburn_moderate, 4)
multi_parallel(Sijtsma2a, c(2, 4))
}
\references{
Cho, E. (2016). Making reliability reliable: A systematic
approach to reliability coefficients. Organizational Research Methods, 19(4),
 651–682.
}
