% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multirel.R
\name{multirel}
\alias{multirel}
\title{Compare multiple multidimensional reliability estimates}
\usage{
multirel(x, until)
}
\arguments{
\item{x}{observed item scores or their covariances}

\item{until}{The number of items up to the first sub-construct}
}
\value{
several multidimensional reliability, omega hiearchical, and subdimensional reliability
}
\description{
This function calculates multidimensional reliability coefficients such as
maximal reliability, correlated factors reliability, stratified alpha,
multidimensional parallel reliability, second-order factor reliability,
bifactor reliability, and related omega hierarchical and sub-dimensional
reliability coefficients.
}
