% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nunnally.R
\name{nunnally}
\alias{nunnally}
\title{Obtain bottom-up approach multidimensional reliability estimates}
\usage{
nunnally(x, until, method = "joreskog", print = TRUE)
}
\arguments{
\item{x}{observed item scores or their covariances}

\item{until}{The number of items up to the first sub-construct}

\item{method}{There are three options. joreskog, mu, kaisercaffrey. mu uses mu4.}

\item{print}{If TRUE, the result is printed to the screen.}
}
\value{
a bifactor reliability estimate
}
\description{
Among several approaches to estimating multidimensional reliability, these
estimators use a bottom-up approach. That is, the test score is divided into 
sub-dimensional or sub-test scores. Multidimensional reliability is 
obtained by estimating the reliability of each subtest score and combining 
them. Different estimates can be obtained depending on how each subtest 
reliability is estimated. These estimators use the general formula first 
proposed by Jum Nunnally.
}
\examples{
nunnally(Osburn_moderate, 4, method = "mu")
nunnally(Sijtsma2a, c(2, 4), method = "kaisercaffrey")
}
\references{
Nunnally, J. C., & Bernstein, I. H. (1994). Psychometric theory 
(3rd ed). McGraw-Hill.
}
\author{
Eunseong Cho, \email{bene@kw.ac.kr}
}
