% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Osburn_moderate-data.R
\docType{data}
\name{Osburn_moderate}
\alias{Osburn_moderate}
\title{Osburn's (2000) eight-component multidimensional data}
\format{
An object of class \code{matrix} (inherits from \code{array}) with 8 rows and 8 columns.
}
\source{
Osburn, H. G. (2000). Coefficient alpha and related internal
consistency reliability coefficients. Psychological Methods, 5(3), 343-355.
https://doi.org/10.1037/1082-989X.5.3.343
}
\usage{
data(Osburn_moderate)
}
\description{
moderate violation of unidimensionality (the average correlation
between sub-constructs is .6)
}
\details{
Osburn's (2000) eight-component multidimensional data
}
\author{
Eunseong Cho, \email{bene@kw.ac.kr}
}
\keyword{datasets}
