% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Peter-data.R
\docType{data}
\name{Peter}
\alias{Peter}
\title{Peter's (1979) actual unidimensional data}
\format{
An object of class \code{matrix} (inherits from \code{array}) with 6 rows and 6 columns.
}
\source{
Peter, J. P., & Ryan, M. J. (1976). An investigation of perceived 
risk at the brand level. Journal of Marketing Research, 13(2), 184-188.
}
\usage{
data(Peter)
}
\description{
Peter's (1979) artificial unidimensional data consisting of six items. This 
dataset originally comes from Peter and Ryan (1976) and measures the 
probability of six types of losses on car purchases.
}
\references{
Peter, J. P. (1979). Reliability: A review of psychometric basics 
and recent marketing practices. Journal of Marketing Research, 16(1), 6-17.
}
\author{
Eunseong Cho, \email{bene@kw.ac.kr}
}
\keyword{datasets}
