% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Raykov-data.R
\docType{data}
\name{Raykov}
\alias{Raykov}
\title{Raykov's (2007) simulated unidimensional data}
\format{
An object of class \code{matrix} (inherits from \code{array}) with 5 rows and 5 columns.
}
\source{
Raykov, T. (2007). Reliability if deleted, not 'alpha if deleted': 
Evaluation of scale reliability following component deletion. British Journal 
of Mathematical and Statistical Psychology, 60(2), 201-216.
}
\usage{
data(Raykov)
}
\description{
Raykov's (2007) simulated unidimensional data consisting of five items.The 
fifth item has 5.6 times greater importance than the other items.
}
\author{
Eunseong Cho, \email{bene@kw.ac.kr}
}
\keyword{datasets}
