% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Reuterberg-data.R
\docType{data}
\name{Reuterberg}
\alias{Reuterberg}
\title{Reuterberg and Gustafsson's (1992) unidimensional data}
\format{
An object of class \code{matrix} (inherits from \code{array}) with 5 rows and 5 columns.
}
\source{
Reuterberg, S.-E., & Gustafsson, J.-E. (1992). Confirmatory factor 
analysis and reliability: Testing measurement model assumptions. Educational 
and Psychological Measurement, 52(4), 795-811.
}
\usage{
data(Reuterberg)
}
\description{
Reuterberg and Gustafsson's (1992) actual dataset consisting of five items. 
This data is the original 40-item verbal test grouped by eight and reduced to 
five items.The CFA reliability coefficient presented by the author is .803.
}
\author{
Eunseong Cho, \email{bene@kw.ac.kr}
}
\keyword{datasets}
