% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/second_order.R
\name{second_order}
\alias{second_order}
\title{Obtain second-order factor reliability estimates of two-dimensional data}
\usage{
second_order(x, until, only_first_grp = FALSE, nonneg_loading = FALSE)
}
\arguments{
\item{x}{observed item scores or their covariances}

\item{until}{The number of items up to the first sub-construct}

\item{only_first_grp}{TRUE if only_first_grp. In this case, this function does not compute a
reliability estimate.}

\item{nonneg_loading}{if TRUE, constraint loadings to nonnegative values}
}
\value{
a second-order factor reliability estimate
}
\description{
Obtain second-order factor reliability estimates of two-dimensional data. It
is a kind of multidimensional CFA reliability, and it is the reliability
derived from the second-order factor model. Items should be grouped by each
sub-dimension. For example, the first dimension consists of the first item
through the until item.
}
\examples{
second_order(Osburn_moderate, 4)
}
\references{
Cho, E. (2016). Making reliability reliable: A systematic
approach to reliability coefficients. Organizational Research Methods, 19(4),
651-682. https://doi.org/10.1177/1094428116656239
}
\author{
Eunseong Cho, \email{bene@kw.ac.kr}
}
