% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Sijtsma2a-data.R
\docType{data}
\name{Sijtsma2a}
\alias{Sijtsma2a}
\title{Sitjsma's (2000) six-component multidimensional data}
\format{
An object of class \code{matrix} (inherits from \code{array}) with 6 rows and 6 columns.
}
\source{
Sijtsma, K. (2009). On the use, the misuse, and the very limited
usefulness of Cronbach’s alpha. Psychometrika, 74(1), 107-120.
https://doi.org/10.1007/s11336-008-9101-0
}
\usage{
data(Sijtsma2a)
}
\description{
Multidimensional data with 3 group factors
}
\details{
Sijtsma's (2009) six-component multidimensional data

A hypothetical data consisting of six items.
}
\author{
Eunseong Cho, \email{bene@kw.ac.kr}
}
\keyword{datasets}
