% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simsek.R
\name{simsek}
\alias{simsek}
\title{Obtain Şimşek-Noyan's theta (principal component analysis reliability)}
\usage{
simsek(x, dim, print = TRUE)
}
\arguments{
\item{x}{a dataframe or a matrix (multidimensional)}

\item{dim}{the number of dimensions}
}
\value{
Şimşek-Noyan's theta
}
\description{
Şimşek-Noyan's (2013) theta is the multidimensional  principal component analysis (PCA)
reliability. It is a multidimensional generalization of Kaiser-Caffrey's alpha.
}
\examples{
Şimşek(Osburn_moderate)
}
\references{
Armor, D. J. (1974). Theta reliability and factor scaling.
In H. L. Costner (Ed.), Sociological methodology (pp. 17-50). Jossey-Bass.

Kaiser, H. F., & Caffrey, J. (1965). Alpha factor analysis.
Psychometrika, 30(1), 1-14.

Şimşek, G. G., & Noyan, F. (2013). McDonald’s ωt, Cronbach’s α,
and Generalized θ for Composite Reliability of Common Factors Structures.
Communications in Statistics - Simulation and Computation, 42(9), 2008–2025.
https://doi.org/10.1080/03610918.2012.689062
}
