% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Spearman-data.R
\docType{data}
\name{Spearman}
\alias{Spearman}
\title{Spearman's (1999)  unidimensional data of six items}
\format{
An object of class \code{matrix} (inherits from \code{array}) with 6 rows and 6 columns.
}
\source{
Spearman, C. (1904). “General intelligence,” objectively determined 
and measured. American Journal of Psychology, 15(2), 201-292
}
\usage{
data(Spearman)
}
\description{
Spearman's (1999)  unidimensional data of six items. The data was collected 
from school children in London. This data is historical data from which the 
first factor analysis was carried out. Spearman's factor analysis model was 
a unidimensional model, but other than this dataset, data that satisfies a 
unidimension were uncommon (Hägglund, 2001).
}
\references{
Hägglund, G. 2001. Milestones in the History of Factor Analysis. 
Pp. 11–38 in Structural equation modeling: Present and future-A festschrift 
in honor of Karl Jöreskog. Scientific Software International.
}
\author{
Eunseong Cho, \email{bene@kw.ac.kr}
}
\keyword{datasets}
