% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/std_alpha.R
\name{std_alpha}
\alias{std_alpha}
\title{Obtain standardized alpha}
\usage{
std_alpha(x)
}
\arguments{
\item{x}{a dataframe or a matrix (unidimensional)}
}
\value{
standardized alpha reliability estimate
}
\description{
Obtain standardized alpha
}
\examples{
std_alpha(Graham1)

}
