% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stratified_alpha.R
\name{stratified_alpha}
\alias{stratified_alpha}
\title{Obtain stratified alpha reliability estimates of two-dimensional data}
\usage{
stratified_alpha(x, until, mp = FALSE)
}
\arguments{
\item{x}{observed item scores or their covariances}

\item{until}{The number of items up to the first sub-construct}

\item{mp}{if TRUE, print multidimensional parallel reliability estimates}
}
\value{
multidimensional_reliability a stratified alpha reliability estimate

subdimensional_reliability reliability estimates of subdimensions

omega_hierarchical general factor saturation
}
\description{
Obtain stratified alpha reliability estimates of two-dimensional data. It
is a multidimensional version of coefficient alpha. Items should be grouped by each
sub-dimension. For example, the first dimension consists of the first item
through the until item.
}
\examples{
stratified_alpha(Osburn_moderate, 4)
stratified_alpha(Sijtsma2a, c(2, 4))
}
\references{
Rajaratnam, N., Cronbach, L. J., & Gleser, G. C. (1965).
Generalizability of stratified-parallel tests. Psychometrika, 30(1), 39–56.
https://doi.org/10.1007/BF02289746

Cho, E. (2016). Making reliability reliable: A systematic approach
to reliability coefficients. Organizational Research Methods, 19(4), 651-682.
https://doi.org/10.1177/1094428116656239
}
\author{
Eunseong Cho, \email{bene@kw.ac.kr}
}
