% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test.tauequivalence.r
\name{test.tauequivalence}
\alias{test.tauequivalence}
\title{Test the essential tau-equivalence of the data}
\usage{
test.tauequivalence(data)
}
\arguments{
\item{data}{a dataframe or a matrix (unidimensional)}
}
\value{
taueq_cfi

taueq_tli

taueq_rmsea

taueq_df

taueq_pvalue

taueq_chisq

conge_cfi

conge_tli

conge_rmsea

conge_df

conge_pvalue

conge_chisq

diff_df

diff_chisq

diff_pvalue
}
\description{
The goodness-of-fit indices are compared with the parameter estimates of the 
essential tau-equivalence model and the cognate model. It can also be used for 
the purpose of investigating the unidimensionality of the data.
}
\examples{
test.tauequivalence(Graham1)
}
\references{
Graham, J. M. (2006). Congeneric and (essentially) tau-equivalent 
estimates of score reliability what they are and how to use them. Educational 
and Psychological Measurement, 66(6), 930-944.

Cho, E. (2016). Making reliability reliable: A systematic 
approach to reliability coefficients. Organizational Research Methods, 19(4), 
651-682.

Cho, E., & Kim, S. (2015). Cronbach’s coefficient alpha: Well 
known but poorly understood. Organizational Research Methods, 18(2), 207-230.
}
\author{
Eunseong Cho, \email{bene@kw.ac.kr}
}
