% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uni_cfa.R
\name{uni_cfa}
\alias{uni_cfa}
\title{Unidimensional confirmatory factor analysis}
\usage{
uni_cfa(
  cov,
  what = "est",
  sample_size = 500,
  nonneg_loading = FALSE,
  taueq = FALSE,
  parallel = FALSE
)
}
\arguments{
\item{cov}{observed covariances}

\item{what}{e.g., "est", "std", "fit"}

\item{sample_size}{number of sample observations}

\item{nonneg_loading}{if TRUE, constraint loadings to nonnegative values}

\item{taueq}{if TRUE, a tau-equivalent model is estimated}

\item{parallel}{if TRUE, a parallel model is estimated}
}
\value{
parameter estimates of unidimensional cfa model
}
\description{
Unidimensional confirmatory factor analysis
}
\examples{
uni_cfa(Graham1)
}
