% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unirel.R
\name{unirel}
\alias{unirel}
\title{Obtain various unidimensional reliablity coefficients}
\usage{
unirel(x, Lambda4.include = TRUE, psych.include = TRUE)
}
\arguments{
\item{x}{a dataframe or a matrix (unidimensional)}

\item{Lambda4.include}{Whether to include the reliability coefficients
provided by the package Lambda4 (lambda5, lambda6, max_lambda, lambda4_max,
lambda4_75)}

\item{psych.include}{Whether to include reliability coefficients
(GLB.algebraic, GLB.fa) provided by the package psych}
}
\description{
R packages that provide unidimensional reliability coefficients include
Lambda4 and psych. These are usually separate functions, so comparing
multiple reliability estimates can take a long time. This function shows the
 reliability coefficients newly added in this package together with the
 reliability coefficients provided by Lambda4 and psych. This function can
 provide Table 2 of Cho(in press) immediately.
}
\examples{
unirel(Graham1)
}
\references{
Cho, E. (in press). Neither Cronbach's alpha nor McDonald's
omega: A comment on Sijtsma and Pfadt. Psychometrika.
}
\seealso{
[Lambda4::lambda5()] for a related function of the package Lambda4

[Lambda4::lambda6()] for a related function of the package Lambda4

[Lambda4::quant.lambda4()] for a related function of the package Lambda4

[psych::GLB.algebraic()] for a related function of the package psych

[psych::GLB.fa()] for a related function of the package psych
}
\author{
Eunseong Cho, \email{bene@kw.ac.kr}
}
