% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Votaw-data.R
\docType{data}
\name{Votaw}
\alias{Votaw}
\title{Votaw's (1948) unidimensional data}
\format{
An object of class \code{matrix} (inherits from \code{array}) with 4 rows and 4 columns.
}
\source{
Votaw, D. F., Jr. Testing compound symmetry in a normal multivariate 
distribution.Annals of Mathematical Statistics, 1948, 19, 447-473.
}
\usage{
data(Votaw)
}
\description{
Votaw's (1948) unidimensional data consisting of four items.
}
\references{
Jöreskog, K. G. (1971). Statistical analysis of sets of 
congeneric tests. Psychometrika, 36(2), 109-133.
}
\author{
Eunseong Cho, \email{bene@kw.ac.kr}
}
\keyword{datasets}
