% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/second_order.R
\name{second_order}
\alias{second_order}
\title{Obtain second-order factor reliability estimates of two-dimensional data}
\usage{
second_order(x, nobs = NULL, until, print = TRUE)
}
\arguments{
\item{x}{observed item scores or their covariances}

\item{nobs}{number of observations (i.e., sample size)}

\item{until}{The number of items up to the first sub-construct}

\item{print}{If TRUE, the result is printed to the screen.}
}
\value{
a second-order factor reliability estimate
}
\description{
Obtain second-order factor reliability estimates of two-dimensional data. It
is a kind of multidimensional CFA reliability, and it is the reliability
derived from the second-order factor model. Items should be grouped by each
sub-dimension. For example, the first dimension consists of the first item
through the until item.
}
\examples{
second_order(Osburn_moderate, 4)
second_order(Sijtsma2a, c(2, 4))
}
\references{
Cho, E. (2016). Making reliability reliable: A systematic
approach to reliability coefficients. Organizational Research Methods, 19(4),
651-682. https://doi.org/10.1177/1094428116656239
}
\author{
Eunseong Cho, \email{bene@kw.ac.kr}
}
